Fixed bugs associated with CompoundKeyRepositoryBase refactoring.
