+++ SharpRepository.Repository/CompoundKeyRepositoryBase.cs
@@ -1,12 +1,9 @@
 using System;
-using System.Collections;
-using System.Reflection;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using SharpRepository.Repository.Caching;
 using SharpRepository.Repository.FetchStrategies;
-using SharpRepository.Repository.Helpers;
 using SharpRepository.Repository.Queries;
 using SharpRepository.Repository.Specifications;
 using SharpRepository.Repository.Transactions;
@@ -69,7 +66,30 @@ namespace SharpRepository.Repository
                 );
         }
 
-        public override T Get(params object[] keys)
+        // These are the actual implementation that the derived class needs to implement
+        protected abstract T GetQuery(params object[] keys);
+
+        public bool Exists(params object[] keys)
+        {
+            return TryGet(out T entity, keys);
+        }
+
+        public bool TryGet(out T entity, params object[] keys)
+        {
+            entity = default(T);
+
+            try
+            {
+                entity = Get(keys);
+                return entity != null;
+            }
+            catch (Exception)
+            {
+                return false;
+            }
+        }
+
+        public T Get(params object[] keys)
         {
             return _queryManager.ExecuteGet(
                 () => GetQuery(keys),
@@ -167,7 +187,16 @@ namespace SharpRepository.Repository
                 _queryManager.OnItemAdded(keys, entity);
         }
 
-        public void Delete(T entity)
+        public void Delete(params object[] keys)
+        {
+            var entity = Get(keys);
+
+            if (entity == null) throw new ArgumentException("No entity exists with these keys.", "keys");
+
+            Delete(entity);
+        }
+
+        public override void Delete(T entity)
         {
             if (entity == null) throw new ArgumentNullException("entity");
 
@@ -191,6 +220,64 @@ namespace SharpRepository.Repository
 
             if (GetPrimaryKeys(entity, out object[] keys))
                 _queryManager.OnItemUpdated(keys, entity);
+        }
+
+        protected virtual bool GetPrimaryKeys(T entity, out object[] keys)
+        {
+            keys = null;
+            var propInfo = GetPrimaryKeyPropertyInfo();
+
+            // if there is no property that matches then return false
+            if (propInfo == null)
+                return false;
+
+            keys = propInfo.Select(info => info.GetValue(entity, null)).ToArray();
+
+            return true;
+        }
+
+        protected virtual bool SetPrimaryKey(T entity, object[] keys)
+        {
+            var propInfo = GetPrimaryKeyPropertyInfo();
+
+            // if there is no property that matches then return false
+            if (propInfo == null || keys == null || propInfo.Length != keys.Length)
+                return false;
+
+            var i = 0;
+            foreach (var key in keys)
+            {
+                propInfo[i].SetValue(entity, key, null);
+                i++;
+            }
+
+            return true;
+        }
+
+        protected virtual ISpecification<T> ByPrimaryKeySpecification(object[] keys)
+        {
+            var propInfo = GetPrimaryKeyPropertyInfo();
+            if (propInfo == null || keys == null || propInfo.Length != keys.Length)
+                return null;
+
+            ISpecification<T> specification = null;
+            var parameter = Expression.Parameter(typeof(T), "x");
+
+            var i = 0;
+            foreach (var lambda in keys.Select(key => Expression.Lambda<Func<T, bool>>(
+                        Expression.Equal(
+                            Expression.PropertyOrField(parameter, propInfo[i].Name),
+                            Expression.Constant(key)
+                        ),
+                        parameter
+                    ))
+                )
+            {
+                specification = specification == null ? new Specification<T>(lambda) : specification.And(lambda);
+                i++;
+            }
+
+            return specification;
         }
 
         private void Save()
@@ -401,7 +488,7 @@ namespace SharpRepository.Repository
                 _queryManager.OnItemAdded(key, key2, entity);
         }
 
-        public void Delete(T entity)
+        public override void Delete(T entity)
         {
             if (entity == null) throw new ArgumentNullException("entity");
 
@@ -701,7 +788,7 @@ namespace SharpRepository.Repository
                 _queryManager.OnItemAdded(key, key2, key3, entity);
         }
 
-        public void Delete(T entity)
+        public override void Delete(T entity)
         {
             if (entity == null) throw new ArgumentNullException("entity");
