+++ SharpRepository.Repository/CompoundKeyRepositoryBaseCommon.cs
@@ -106,32 +106,7 @@ namespace SharpRepository.Repository
         public abstract IRepositoryQueryable<TResult> Join<TJoinKey, TInner, TResult>(IRepositoryQueryable<TInner> innerRepository, Expression<Func<T, TJoinKey>> outerKeySelector, Expression<Func<TInner, TJoinKey>> innerKeySelector, Expression<Func<T, TInner, TResult>> resultSelector)
             where TInner : class
             where TResult : class;
-
-        // These are the actual implementation that the derived class needs to implement
-        protected abstract T GetQuery(params object[] keys);
-
-        public bool Exists(params object[] keys)
-        {
-            return TryGet(out T entity, keys);
-        }
-
-        public bool TryGet(out T entity, params object[] keys)
-        {
-            entity = default(T);
-
-            try
-            {
-                entity = Get(keys);
-                return entity != null;
-            }
-            catch (Exception)
-            {
-                return false;
-            }
-        }
-
-        public abstract T Get(params object[] keys);
-
+        
         // These are the actual implementation that the derived class needs to implement
         protected abstract IQueryable<T> FindAllQuery(ISpecification<T> criteria);
         protected abstract IQueryable<T> FindAllQuery(ISpecification<T> criteria, IQueryOptions<T> queryOptions);
@@ -285,21 +260,14 @@ namespace SharpRepository.Repository
         // This is the actual implementation that the derived class needs to implement
         protected abstract void DeleteItem(T entity);
 
+        public abstract void Delete(T entity);
+
         public void Delete(IEnumerable<T> entities)
         {
             foreach (var entity in entities)
             {
                 Delete(entity);
             }
-        }
-
-        public void Delete(params object[] keys)
-        {
-            var entity = Get(keys);
-
-            if (entity == null) throw new ArgumentException("No entity exists with these keys.", "keys");
-
-            Delete(entity);
         }
 
         public void Delete(Expression<Func<T, bool>> predicate)
@@ -330,64 +298,6 @@ namespace SharpRepository.Repository
         protected abstract void SaveChanges();
 
         public abstract void Dispose();
-
-        protected virtual bool GetPrimaryKeys(T entity, out object[] keys)
-        {
-            keys = null;
-            var propInfo = GetPrimaryKeyPropertyInfo();
-
-            // if there is no property that matches then return false
-            if (propInfo == null)
-                return false;
-
-            keys = propInfo.Select(info => info.GetValue(entity, null)).ToArray();
-
-            return true;
-        }
-
-        protected virtual bool SetPrimaryKey(T entity, object[] keys)
-        {
-            var propInfo = GetPrimaryKeyPropertyInfo();
-
-            // if there is no property that matches then return false
-            if (propInfo == null || keys == null || propInfo.Length != keys.Length)
-                return false;
-
-            var i = 0;
-            foreach (var key in keys)
-            {
-                propInfo[i].SetValue(entity, key, null);
-                i++;
-            }
-
-            return true;
-        }
-
-        protected virtual ISpecification<T> ByPrimaryKeySpecification(object[] keys)
-        {
-            var propInfo = GetPrimaryKeyPropertyInfo();
-            if (propInfo == null || keys == null || propInfo.Length != keys.Length)
-                return null;
-
-            ISpecification<T> specification = null;
-            var parameter = Expression.Parameter(typeof(T), "x");
-
-            var i = 0;
-            foreach (var lambda in keys.Select(key => Expression.Lambda<Func<T, bool>>(
-                        Expression.Equal(
-                            Expression.PropertyOrField(parameter, propInfo[i].Name),
-                            Expression.Constant(key)
-                        ),
-                        parameter
-                    ))
-                )
-            {
-                specification = specification == null ? new Specification<T>(lambda) : specification.And(lambda);
-                i++;
-            }
-
-            return specification;
-        }
 
         protected virtual PropertyInfo[] GetPrimaryKeyPropertyInfo()
         {
